feat: change name, delete console log, delete area from PhoneInput
