+++ packages/web/src/common/components/Forms/PhoneInput.tsx
@@ -1,21 +1,116 @@
-import React, { ChangeEvent, useEffect, useState } from "react";
-import TextInput, {
-  TextInputProps,
-} from "@sparcs-students/web/common/components/Forms/TextInput";
+/*
+--- 사용 방법 (예시) ---
+<NewPhoneInput
+  label="전화번호"
+  placeholder="010-XXXX-XXXX"
+  value={phone}
+  handleChange={setPhone}
+  setErrorStatus={setErrorPhone}
+/>
+*/
 
-interface PhoneInputProps extends Omit<TextInputProps, "onChange"> {
-  value: string;
-  onChange: (value: string) => void;
+import React, {
+  ChangeEvent,
+  ChangeEventHandler,
+  InputHTMLAttributes,
+  useEffect,
+  useLayoutEffect,
+  useRef,
+  useState,
+} from "react";
+
+import isPropValid from "@emotion/is-prop-valid";
+import styled, { css } from "styled-components";
+
+import ErrorMessage from "./_atomic/ErrorMessage";
+import Label from "./_atomic/Label";
+
+export interface PhoneInputProps
+  extends InputHTMLAttributes<HTMLInputElement | HTMLTextAreaElement> {
+  label?: string;
+  placeholder: string;
+  disabled?: boolean;
+  value?: string;
+  handleChange?: (value: string) => void;
+  setErrorStatus?: (hasError: boolean) => void;
+  onChange?: ChangeEventHandler<HTMLInputElement>;
 }
 
+const errorBorderStyle = css`
+  border-color: ${({ theme }) => theme.colors.RED[600]};
+`;
+
+const disabledStyle = css`
+  background-color: ${({ theme }) => theme.colors.GRAY[100]};
+  border-color: ${({ theme }) => theme.colors.GRAY[200]};
+`;
+
+const Input = styled.input.withConfig({
+  shouldForwardProp: prop => isPropValid(prop) && prop !== "hasError",
+})<PhoneInputProps & { hasError: boolean }>`
+  display: block;
+  width: 100%;
+  padding: 8px 12px 8px 12px;
+  outline: none;
+  border: 1px solid ${({ theme }) => theme.colors.GRAY[200]};
+  border-radius: 4px;
+  gap: 8px;
+  font-family: ${({ theme }) => theme.fonts.FAMILY.PRETENDARD};
+  font-size: 16px;
+  line-height: 20px;
+  font-weight: ${({ theme }) => theme.fonts.WEIGHT.REGULAR};
+  color: ${({ theme }) => theme.colors.BLACK};
+  background-color: ${({ theme }) => theme.colors.WHITE};
+  &:focus {
+    border-color: ${({ theme, hasError, disabled }) =>
+      !hasError && !disabled && theme.colors.PRIMARY};
+  }
+  &:hover:not(:focus) {
+    border-color: ${({ theme, hasError, disabled }) =>
+      !hasError && !disabled && theme.colors.GRAY[300]};
+  }
+  &::placeholder {
+    color: ${({ theme }) => theme.colors.GRAY[200]};
+  }
+  ${({ disabled }) => disabled && disabledStyle}
+  ${({ hasError }) => hasError && errorBorderStyle}
+`;
+
+const InputWrapper = styled.div`
+  width: 100%;
+  flex-direction: column;
+  display: flex;
+  gap: 4px;
+`;
+
+// Component
 const PhoneInput: React.FC<PhoneInputProps> = ({
   label = "",
+  placeholder,
+  disabled = false,
   value = "",
-  onChange = () => {},
+  handleChange = () => {}, // setValue
+  setErrorStatus = () => {},
+  onChange = undefined, // display results (complicated)
   ...props
 }) => {
   const [error, setError] = useState("");
   const [touched, setTouched] = useState(false);
+  const inputRef = useRef<HTMLInputElement | null>(null);
+  const cursorRef = useRef<number>(0);
+
+  useLayoutEffect(() => {
+    if (inputRef.current) {
+      inputRef.current.setSelectionRange(cursorRef.current, cursorRef.current);
+    }
+  }, [value]);
+
+  useEffect(() => {
+    const hasError = !!error;
+    if (setErrorStatus) {
+      setErrorStatus(hasError);
+    }
+  }, [error, setErrorStatus]);
 
   useEffect(() => {
     if (touched) {
@@ -27,7 +122,10 @@ const PhoneInput: React.FC<PhoneInputProps> = ({
         setError("필수로 채워야 하는 항목입니다");
       } else if (!isValidFormat) {
         setError("숫자만 입력 가능합니다");
-      } else if (value.replace(/-/g, "").length !== 11) {
+      } else if (
+        value.replace(/-/g, "").length !== 11 ||
+        value.slice(0, 3) !== "010"
+      ) {
         setError("유효하지 않은 전화번호입니다");
       } else {
         setError("");
@@ -37,16 +135,6 @@ const PhoneInput: React.FC<PhoneInputProps> = ({
 
   const handleBlur = () => {
     setTouched(true);
-  };
-
-  const handleChange = (
-    e: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>,
-  ) => {
-    const inputValue = e.target.value;
-
-    if (inputValue.length <= 13) {
-      onChange(inputValue);
-    }
   };
 
   const formatValue = (nums: string) => {
@@ -64,17 +152,48 @@ const PhoneInput: React.FC<PhoneInputProps> = ({
     return formattedInput;
   };
 
+  const handlePhoneValueChange = (e: ChangeEvent<HTMLInputElement>) => {
+    const inputValue = e.target.value;
+    const currentCursor = inputRef.current?.selectionStart || 0;
+    const lengthDifference = inputValue.length - formatValue(value).length;
+    if (
+      lengthDifference > 0 &&
+      (currentCursor === 3 ||
+        currentCursor === 4 ||
+        currentCursor === 8 ||
+        currentCursor === 9)
+    ) {
+      cursorRef.current = currentCursor + 1;
+    } else if (
+      lengthDifference < 0 &&
+      ((currentCursor === 4 && inputValue.length === 4) ||
+        (currentCursor === 9 && inputValue.length === 9))
+    ) {
+      cursorRef.current = currentCursor - 1;
+    } else {
+      cursorRef.current = currentCursor;
+    }
+    if (inputValue.length <= 13) handleChange(inputValue);
+  };
+
   return (
-    <TextInput
-      label={label}
-      value={formatValue(value)}
-      onChange={handleChange}
-      errorMessage={error}
-      onBlur={handleBlur}
-      {...props}
-    />
+    <InputWrapper>
+      {label && <Label>{label}</Label>}
+      <InputWrapper>
+        <Input
+          placeholder={placeholder}
+          hasError={!!error}
+          disabled={disabled}
+          value={formatValue(value)}
+          onChange={onChange ?? handlePhoneValueChange}
+          onBlur={handleBlur}
+          ref={inputRef}
+          {...props}
+        />
+        {error && <ErrorMessage>{error}</ErrorMessage>}
+      </InputWrapper>
+    </InputWrapper>
   );
 };
-// TODO: DB 값 default로 넣어두기
 
 export default PhoneInput;
